Renaming MNLDiscreteChoiceStep -> LargeMultinomialLogitStep
