+++ urbansim_templates/models/__init__.py
@@ -1,5 +1,5 @@
 from .binary_logit import BinaryLogitStep
-from .dcm import MNLDiscreteChoiceStep
+from .large_multinomial_logit import LargeMultinomialLogitStep
 from .regression import OLSRegressionStep
 from .shared import TemplateStep
 from .small_multinomial_logit import SmallMultinomialLogitStep
